backend/flask_api/requirements.txt
Flask
tensorflow
Pillow
requests
google-generativeai
python-dotenv
flask-cors
pytest